things/pyproject.toml
[tool.poetry]
name = "drf_api_permissions_postgres"
version = "0.1.0"
description = ""
authors = ["mhmad wrekat <mhmmadwreekkat@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.12.4"
pytz = "^2021.3"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
